lab2/requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.9  # Esta dependencia es para la conexión con PostgreSQL
python-dotenv==1.0.1